+++ eternalcore-api/src/main/java/com/eternalcode/core/feature/afk/event/AfkSwitchEvent.java
@@ -13,15 +13,21 @@ public class AfkSwitchEvent extends Event implements Cancellable {
     private static final HandlerList HANDLER_LIST = new HandlerList();
 
     private final Afk afk;
+    private final boolean isAfk;
     private boolean cancelled;
 
-    public AfkSwitchEvent(Afk afk) {
+    public AfkSwitchEvent(Afk afk, boolean isAfk) {
         super(false);
         this.afk = afk;
+        this.isAfk = isAfk;
     }
 
     public Afk getAfk() {
         return this.afk;
+    }
+
+    public boolean isAfk() {
+        return this.isAfk;
     }
 
     @Override
