+++ eternalcore-core/src/main/java/com/eternalcode/core/feature/afk/AfkKickController.java
@@ -9,52 +9,49 @@ import com.eternalcode.core.translation.Translation;
 import com.eternalcode.core.translation.TranslationManager;
 import com.eternalcode.core.user.User;
 import com.eternalcode.core.user.UserManager;
-import com.eternalcode.core.viewer.Viewer;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.minimessage.MiniMessage;
-import org.bukkit.Bukkit;
+import org.bukkit.Server;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
-import org.bukkit.event.player.PlayerMoveEvent;
-import org.bukkit.event.player.PlayerQuitEvent;
 
 import java.util.UUID;
 
 @Controller
-class AfkController implements Listener {
+public class AfkKickController implements Listener {
 
-    private final AfkService afkService;
     private final MiniMessage miniMessage;
+    private final Server server;
     private final UserManager userManager;
+    private final PluginConfiguration configuration;
     private final TranslationManager translationManager;
-    private final PluginConfiguration config;
 
     @Inject
-    AfkController(
-    AfkService afkService,
-    MiniMessage miniMessage,
-    UserManager userManager,
-    TranslationManager translationManager, 
-    PluginConfiguration config
-   ) {
-        this.afkService = afkService;
+    public AfkKickController(
+        MiniMessage miniMessage,
+        Server server,
+        UserManager userManager,
+        PluginConfiguration configuration,
+        TranslationManager translationManager
+    ) {
         this.miniMessage = miniMessage;
+        this.server = server;
         this.userManager = userManager;
+        this.configuration = configuration;
         this.translationManager = translationManager;
-        this.config = config;
     }
 
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
     void onAfkSwitch(AfkSwitchEvent event) {
         UUID playerUUID = event.getAfk().getPlayer();
 
-        if (!this.afkService.isAfk(playerUUID) || !this.config.afk.kickOnAfk) {
+        if (!event.isAfk() || !this.configuration.afk.kickOnAfk) {
             return;
         }
 
-        Player player = Bukkit.getPlayer(playerUUID);
+        Player player = this.server.getPlayer(playerUUID);
 
         if (player == null) {
             return;
@@ -66,17 +63,4 @@ class AfkController implements Listener {
         Component component = this.miniMessage.deserialize(translation.afk().afkKickReason());
         player.kickPlayer(Legacy.SECTION_SERIALIZER.serialize(component));
     }
-
-    @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
-    void onMove(PlayerMoveEvent event) {
-        UUID uniqueId = event.getPlayer().getUniqueId();
-
-        this.afkService.markInteraction(uniqueId);
-    }
-
-    @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
-    void onQuit(PlayerQuitEvent event) {
-        this.afkService.clearAfk(event.getPlayer().getUniqueId());
-    }
-
 }
